Relocate framework test files
