+++ test/beta/observer/test_observer.py
@@ -2,10 +2,13 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+import asyncio
+import logging
 from contextlib import ExitStack
 
 import pytest
 
+from autogen.beta.annotations import Context as AnnContext
 from autogen.beta.context import ConversationContext as Context
 from autogen.beta.events import ModelMessage, ObserverAlert, Severity, ToolCallEvent
 from autogen.beta.observer import BaseObserver
@@ -104,3 +107,57 @@ class TestBaseObserver:
 
             await stream.send(ToolCallEvent(name="t", arguments="{}"), ctx)
             assert obs.process_count == 1
+
+
+class _CrashingObserver(BaseObserver):
+    """Observer whose ``process()`` always raises."""
+
+    def __init__(self) -> None:
+        super().__init__("crasher", watch=EventWatch(ModelMessage))
+
+    async def process(self, events, ctx):
+        raise RuntimeError("observer exploded")
+
+
+class _NullModelMessageObserver(BaseObserver):
+    def __init__(self) -> None:
+        super().__init__("null", watch=EventWatch(ModelMessage))
+
+    async def process(self, events, ctx):
+        return None
+
+
+@pytest.mark.asyncio
+class TestObserverExceptionHandling:
+    async def test_observer_process_exception_is_caught(self, caplog) -> None:
+        observer = _CrashingObserver()
+        stream = MemoryStream()
+        ctx = Context(stream=stream)
+
+        with ExitStack() as stack:
+            observer.register(stack, ctx)
+
+            with caplog.at_level(logging.ERROR):
+                await stream.send(ModelMessage(content="trigger"), ctx)
+                await asyncio.sleep(0.01)
+
+        assert any("process() failed" in r.message for r in caplog.records)
+
+    async def test_observer_returns_none_no_signal(self) -> None:
+        observer = _NullModelMessageObserver()
+        stream = MemoryStream()
+        ctx = Context(stream=stream)
+
+        signals: list[ObserverAlert] = []
+
+        async def _capture(event: ObserverAlert, _ctx: AnnContext) -> None:
+            signals.append(event)
+
+        stream.where(ObserverAlert).subscribe(_capture)
+
+        with ExitStack() as stack:
+            observer.register(stack, ctx)
+            await stream.send(ModelMessage(content="test"), ctx)
+            await asyncio.sleep(0.01)
+
+        assert len(signals) == 0
