+++ test/beta/test_compact.py
