+++ test/beta/test_watch.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import asyncio
+import datetime
 
 import pytest
 
@@ -13,6 +14,7 @@ from autogen.beta.watch import (
     AllOf,
     AnyOf,
     CadenceWatch,
+    CronWatch,
     DelayWatch,
     EventWatch,
     IntervalWatch,
@@ -506,3 +508,80 @@ class TestSequence:
         await stream.send(ToolCallEvent(name="t2", arguments="{}"), ctx)
         await stream.send(ModelMessage(content="m2"), ctx)
         assert len(received) == 2
+
+
+class TestCronWatchExpressions:
+    def test_range_expression(self) -> None:
+        cron = CronWatch("1-5 * * * *")
+        now = datetime.datetime(2026, 3, 22, 10, 0, 0)
+        next_fire = cron._next_fire_time(now)
+        assert next_fire.minute in {1, 2, 3, 4, 5}
+
+    def test_list_expression(self) -> None:
+        cron = CronWatch("0,15,30,45 * * * *")
+        now = datetime.datetime(2026, 3, 22, 10, 0, 0)
+        next_fire = cron._next_fire_time(now)
+        assert next_fire.minute in {0, 15, 30, 45}
+
+    def test_step_with_range(self) -> None:
+        cron = CronWatch("*/10 * * * *")
+        now = datetime.datetime(2026, 3, 22, 10, 0, 0)
+        next_fire = cron._next_fire_time(now)
+        assert next_fire.minute in {0, 10, 20, 30, 40, 50}
+
+    def test_specific_hour_and_minute(self) -> None:
+        cron = CronWatch("30 14 * * *")
+        now = datetime.datetime(2026, 3, 22, 10, 0, 0)
+        next_fire = cron._next_fire_time(now)
+        assert next_fire.hour == 14
+        assert next_fire.minute == 30
+
+    def test_invalid_field_count_raises(self) -> None:
+        cron = CronWatch("* * *")
+        with pytest.raises(ValueError, match="5 fields"):
+            cron._next_fire_time(datetime.datetime.now())
+
+    def test_step_five_minutes(self) -> None:
+        cron = CronWatch("*/5 * * * *")
+        now = datetime.datetime(2026, 3, 21, 10, 3, 0)
+        nxt = cron._next_fire_time(now)
+        assert nxt.minute == 5
+        assert nxt.hour == 10
+
+    def test_day_of_week_name(self) -> None:
+        cron = CronWatch("0 9 * * MON")
+        # 2026-03-21 is a Saturday; next Monday is 2026-03-23
+        now = datetime.datetime(2026, 3, 21, 10, 0, 0)
+        nxt = cron._next_fire_time(now)
+        assert nxt.weekday() == 0
+        assert nxt.hour == 9
+        assert nxt.minute == 0
+
+    def test_invalid_expression_raises(self) -> None:
+        cron = CronWatch("bad")
+        with pytest.raises(ValueError, match="Invalid cron"):
+            cron._next_fire_time(datetime.datetime.now())
+
+    def test_numeric_dow_sunday_zero(self) -> None:
+        cron = CronWatch("0 9 * * 0")
+        # 2026-03-21 is a Saturday; next Sunday is 2026-03-22
+        now = datetime.datetime(2026, 3, 21, 10, 0, 0)
+        nxt = cron._next_fire_time(now)
+        assert nxt.isoweekday() % 7 == 0
+        assert nxt.day == 22
+        assert nxt.hour == 9
+
+    def test_numeric_dow_saturday_six(self) -> None:
+        cron = CronWatch("0 9 * * 6")
+        # 2026-03-21 (Saturday) at 10:00 — past 9:00; next Saturday is 2026-03-28
+        now = datetime.datetime(2026, 3, 21, 10, 0, 0)
+        nxt = cron._next_fire_time(now)
+        assert nxt.isoweekday() % 7 == 6
+        assert nxt.day == 28
+        assert nxt.hour == 9
+
+    def test_numeric_dow_seven_is_sunday_alias(self) -> None:
+        watch_seven = CronWatch("0 9 * * 7")
+        watch_zero = CronWatch("0 9 * * 0")
+        now = datetime.datetime(2026, 3, 21, 10, 0, 0)
+        assert watch_seven._next_fire_time(now) == watch_zero._next_fire_time(now)
